requirements.txt
PyYAML==3.11
click==6.2
